Fix issuues #58, #59, #60
