+++ lib/util/author.py
@@ -3,11 +3,13 @@ import re
 from util.read_utils import lines_per_n
 
 
-def write_author_uid_map():
+def get_uid_map(write_to_file=True):
     """
     This function is used to generate and write to a JSON file the mapping of authors to a unique integer identifier.
     Authors are identified through a regular expression search for their email addresses. The integer identifiers
     generated are used in other modules like the generation and statistical analysis of hyperedges.
+    :param write_to_file: If true, results are written to author_uid_map.json (default=True)
+    :return: A list of all message ids that are leaf nodes
     """
     index = 0
     author_set = set()
@@ -17,13 +19,11 @@ def write_author_uid_map():
     with open('clean_data.json', 'r') as json_file:
         for chunk in lines_per_n(json_file, 9):
             json_obj = json.loads(chunk)
-            # print("\nFrom", json_obj['From'], "\nTo", json_obj['To'], "\nCc", json_obj['Cc'])
             from_addr = email_re.search(json_obj['From'])
             author_set.add(from_addr.group(0) if from_addr is not None else json_obj['From'])
             author_set |= set(email_re.findall(json_obj['To']))
             if json_obj['Cc'] is not None:
                 author_set |= set(email_re.findall(json_obj['Cc']))
-            # print("\nFrom", json_obj['From'], "\nTo", json_obj['To'], "\nCc", json_obj['Cc'])
     print("JSON data loaded.")
 
     for address in author_set:
@@ -34,3 +34,4 @@ def write_author_uid_map():
         json.dump(author_uid_map, map_file, indent=1)
         map_file.close()
     print("UID map written to author_uid_map.json.")
+    return author_uid_map
